filetype: fix OOB read when detecting type of truncated kernel images
The ARM and RISC-V kernel Image format features a 64-byte header and
places the 32-bit magic value identifying it at offset 56.

The check for the magic values should thus ensure that at least 56 bytes
are guaranteed to be available in the buffer, thus move it into
the >= 64 byte segment of the function.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240517074735.766656-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>