usbgadget: split usbgadget_register into prepare and register
Currently usbgadget_register() prepares the struct::f_multi_opts
according a given configuration string and registers the the usbgadget
device.

This is not optimal for adding custom fastboot_opts. Split the single
function into usbgadget_prepare() and usbgadget_register() to avoid code
duplication. This way it's easy to reuse the preparation code and add
the custom fastboot_opts afterwards.

In addition usbgadget_prepare_register() is added to keep the simplicity
of the single prepare and register functionality.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20240521080617.1532603-2-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>