watchdog: fix watchdog restart handler when autoping is enabled
When the watchdog autoping feature is enabled it will ping the watchdog
every 0.5s. This becomes a problem when the fallback restart handler is
called to reset the CPU. It will then set the watchdog to trigger in one
second and waits for CPU reset using mdelay(). During the mdelay() the
autoping poller will continue to feed the watchdog, so we never actually
reset the CPU.

Use mdelay_non_interruptible() instead so that no poller can run in the
background.

Link: https://lore.barebox.org/20240527095057.3418435-2-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>