commands: mipi_dbi: fix null pointer dereference when writing data
When accessing argv[optind + i], the whole array index must stay below
argc, not only i. Also val is only 4 bytes long, so when indexed with i
(which was initialised to optind, which is at least 3) will overflow
after reading one data argument from the command line. Add a guard
against the latter case, and initialise i to 0 to fix the first problem.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.barebox.org/20240527132842.1016597-1-rhi@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>