net: smc911x: Fix amount of words read from RX FIFO
According to the datasheet the packet length field contains the packet
length in bytes (and it's also used like this in the driver). This means
the number of dwords to read from the FIFO is one quarter of that, or,
if we account for non dword aligned packet lengths, as:

	len = (pktlen + 3) / 4.

I have no idea why the code has an additional + 2 when calculating the
number of dwords. The corresponding U-Boot driver also does not have
this additional + 2, so drop it from the barebox driver as well.

Link: https://lore.barebox.org/20240528085806.1172689-7-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>