ARM: arria10: reset-manager: don't touch l4 wdog
Leave the L4 watchdog 0 running.
If a watchdog reset happens while booting, the bootrom attempts to load
the last valid bootloader identified by the initswlastld register.

As we now hang() in arria10-xload after 10 failed FPGA writing attempts,
we need the watchdog to trigger so that the bootrom can try to start the
next bootloader image as fallback.

Signed-off-by: Steffen Trumtrar <s.trumtrar@pengutronix.de>
Link: https://lore.barebox.org/20240617-v2024-05-0-topic-socfpga-arria10-xloader-v1-4-48a6eba4bb5e@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>