ARM: arria10: xload: fix __arria10_load_fpga
If arria10_load_fpga is called with an offset the (count < size)
comparison must handle that; increasing the count variable doesn't make
the bitstream size smaller.
As the function actually works with sectors, rename the variables to
make that clearer and add the offset to the bitstream_size to get the
correct end_sector up until __arria10_load_fpga needs to read.

Signed-off-by: Steffen Trumtrar <s.trumtrar@pengutronix.de>
Link: https://lore.barebox.org/20240617-v2024-05-0-topic-socfpga-arria10-xloader-v1-7-48a6eba4bb5e@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>