ARM: arria10-xload: rework broken bitstream handling
When after 10 tries, the bitstream still couldn't be loaded, mark the
boot as failed in ARRIA10_SYSMGR_ROM_INITSWSTATE.
This information can be retrieved on barebox start (in lowlevel.c) to
decide what to do.

The bootROM will *not* boot the next boot slot, but will just try again
booting the failed bootloader slot.

Signed-off-by: Steffen Trumtrar <s.trumtrar@pengutronix.de>
Link: https://lore.barebox.org/20240617-v2024-05-0-topic-socfpga-arria10-xloader-v1-14-48a6eba4bb5e@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>