mci: core: remove dev_add_param_bool error paths
dev_add_param_bool fails when the device already has a parameter with
the same name or if there's no memory. The first error can't happen here
and the second won't go unnoticed anyway as something critical is bound
to panic.

Remove thus the error paths, which are never triggered.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240617114142.1546941-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>