efi: payload: inform users on attempt to boot non EFI-stubbed kernel
Calling bootm on an x86 kernel image that lacks an EFI stub will result
in a "no image handler found for image type MBR sector", which can be
very confusing.

Improve upon this by directly suggesting that the kernel's
CONFIG_EFI_STUB option needs to be enabled, so the image has the
expected PE magic (and the EFI stub code that comes with it).

Cc: cybin <info@arachnodroid.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Tested-by: Michael Taubert <info@arachnodroid.de>
Link: https://lore.barebox.org/20240620071616.3228015-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>