common: misc: do not run pollers inside panic()
When panic() is not configured to hang, it will wait 100 ms for messages
to go out before trying to poweroff or reset the machine.

The wait is implemented using udelay(), which runs pollers if the delay
duration exceeds 100 us. This is a bad idea as it needlessly exposes
barebox to hang indefinitely if the driver that's being polled had
something to do with why we ended up in the panic in the first place.

Switch to a non_interruptible variant to busy loop without invoking
pollers.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240620125829.550300-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>