video: backlight: make brightness and slew time parameters unsigned
Despite the backlight parameter being registered as uint32,
backlight_brightness_set took the brightness level as a signed value.

This can lead to wrap around for very large values of brightness
supplied as device parameter and if this results in a negative
brightness, the function appears to hang, because the slew rate loop
runs for a long time.

Fix this by making brightness level an unsigned entity.

Fixes: e3ea90860e99 ("video: Add backlight support")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240701072714.163239-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>