pinctrl: import <linux/pinctrl/pinconf-generic.h> header
The header is an integral part of the Linux pinctrl support, because
pinctrl drivers parse the device tree into the packed format described
by it and receive the same format later when a pinctrl state needs to be
selected.

We don't follow this same procedure in barebox and instead parse pinctrl
entries on demand, but nevertheless, drivers may use these definitions for
their own purposes internally, so let's import the header sans external
definitions.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240701073220.165946-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>