bitops: change test_bit return type from int to bool
test_bit currently returns some positive integer on success and don't
enforce the return values to be either false or true.

This can lead to problems if the return value is not used for
truthiness, but in arithmetic.

This seems to be the case only in KASAN code, where it's used to count
errors. Change the return type to fix this.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240701073509.176191-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>