ARM: fix OP-TEE DT reservation clashing with barebox area at endmem
The barebox area reservation is meant to fix the issue of having gaps
between early barebox-related reservations that might be clobbered by
later calls to request_sdram_region().

The OP-TEE memory region, if it exists, is not a barebox-related
reservation and treating it as such will clash if the reservation is
also listed in the device-tree:

  WARNING: of-reserved-mem: couldn't request reserved sdram region 0xdd000000-0xdfffffff

This failed reservation can lead to a failure to boot, e.g. on
STM32MP13:

  WARNING: of-reserved-mem: couldn't request reserved sdram region 0xdd000000-0xdfffffff
  E/TC:0   tzc_it_handler:26 TZC permission failure
  E/TC:0   dump_fail_filter:420 Permission violation on filter 0
  E/TC:0   dump_fail_filter:425 Violation @0xde03c530, non-secure privileged read, AXI ID 5c0
  E/TC:0   Panic at core/arch/arm/plat-stm32mp1/plat_tzc400.c:30 <tzc_it_handler>
  E/TC:0   TEE load address @ 0xde000000
  E/TC:0   Call stack:

Fix this by defining the end of the barebox area to not include any
OP-TEE memory regions.

Fixes: 3f7b6146669c ("ARM: cpu: start: register barebox memory area")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240701100859.3876429-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>