sandbox: use fixed-layout binding
Device trees used with barebox are likely to still use the
deprecated nvmem cell binding. To have some coverage of the new binding,
switch sandbox over to it.

No functional change.

Note that the nvmem-cells compatible node seems to be no longer
necessary per binding, but that aspect is not yet supported by barebox.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240703190937.689547-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>