lib: kasan: kasan_test: hide buggy access after xzalloc from compiler
Commit 187b85361236 ("KASan: test_kasan: hide buggy accesses from
compiler") used OPTIMIZER_HIDE_VAR() to hide the origin of pointers.

This became necessary when __alloc attributes were added to malloc().
We now add the same attributes to xzalloc as well bring back warnings
for the single xzalloc call in the file.

That memory violation, like the ones before it, is intended as
test_kasan is meant to trigger kasan at runtime to verify correct
operation.

Therefore use OPTIMIZER_HIDE_VAR here too to silence the warning.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240722162635.3828172-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>