resource: return NULL for error in request_barebox_region
The function currently returns NULL for error if the range is in SDRAM
and couldn't be requested or an error pointer if the range is outside.

Reduce the confusion by using only one way to indicate error. As
request_barebox_region is used to replace request_sdram_region calls and
that returns NULL on error, follow suit.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240722170606.3866957-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>