dma: debug: fix compiler warning about pointer u64 cast
All other prints in the file use the DMA address except for this single
print inside debug_dma_map(). Let's use the DMA address there as well to
fix the warning about casting a 32-bit pointer to a 64 bit integer and
to align error messages in case of non-1:1 mapped DMA addresses.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240724094732.886245-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>