usb: gadget: fsl_udc: dma_map buffers instead of using dma_sync
CONFIG_DMA_API_DEBUG depends on DMA buffers being mapped with
dma_map_single before ownership is switched with dma_sync_single.

As this is currently not done, CONFIG_DMA_API_DEBUG reports a lot of
false positives when the driver is enabled. Fix it by using
usb_gadget_map_request which internally calls dma_map_single.

This fixes the false positives, aligns us with what the Linux driver
is doing and reduces overhead when unmapping the buffer for an IN
target endpoint:

Previously with DMA_BIDIRECTIONAL, we always invalidated buffers
when taking back ownership for the CPU. This is unnecessary when
unmapping a buffer for an IN endpoint as the direction is DMA_TO_DEVICE
and we won't be reading back the result.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240724094732.886245-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>