calloc: handle wrap around in total buffer size correctly
calloc() should allocate a memory buffer that fits the product of its
arguments or return NULL if this is not possible.

We violated this so far and a wraparound would result in allocating a
too small buffer leading to buffer overflows.

Fix this by using size_mull which saturates at SIZE_MAX, which malloc
should gracefully return NULL for.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240724094744.894205-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>