env: suggest global/nv command when calling setenv on non-existent parameter
Unlike regular environment variables, device parameters have to be
explicitly created before they are used:

  global.linux.bootargs.memsize="mem=3072M"
  Cannot set parameter global.linux.bootargs.memsize: No such device

The common remedy is to remove the first dot to call the global command
instead that will create the device parameter and then set it:

  global linux.bootargs.memsize="mem=3072M"

This is a common source of confusion and new users trip over it so
often, that we should do something about it.

Initially I wanted the setting of linux.bootargs.* linux.mtdparts.*
and linux.blkdevparts.* to automatically create the variable.
This is easier said than done though and would need a rework of existing
code as well as violation of the current layering.

Instead, we take the easy way out and just suggest the user that the
user create the variable in question if it's not there yet:

  Parameter not found. Did you mean to create it with: global linux.bootargs.memsize='mem=3072M' ?

While at it, we also replace strerror() with %pe.

Reported-by: Sven Püschel <s.pueschel@pengutronix.de>
Reported-by: Casey Reeves <contact@xogium.me>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240726112556.3548806-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>