fs: give erase() a new erase_type parameter
Many managed flashes provide an erase operation, which need not be used
for regular writes, but is provided nonetheless for efficient clearing
of large regions of storage to a fixed bit pattern. The erase type allows
users to specify _why_ the erase operation is done, so the driver or
the core code can choose the most optimal operation.

Existing code is annotated with either ERASE_TO_WRITE for code that is
immediately followed by a write and ERASE_TO_CLEAR for the standalone
erase operations of the shell and fastboot erase command.

Use of an enum allows the future addition of more types, e.g.:

  ERASE_TO_ALL_ZERO: to efficiently implement android sparse all-ones
                     (like util-linux blkdiscard --zeroout)
  ERASE_TO_ALL_ONE: to efficiently implement android sparse all-ones
  ERASE_TO_DISCARD: as alternative to ERASE_TO_CLEAR, when unwritten
                    regions are indeed allowed to have arbitrary content
                     (like util-linux blkdiscard without argument)
  ERASE_TO_SECURE: erase data in a manner appropriate for wiping secrets
                     (like util-linux blkdiscard --secure)

We don't introduce any functional change here yet, as no device yet sets
DEVFS_WRITE_AUTOERASE. This will follow in a separate commit.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240731080510.364706-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>