commands: sync: add new command to flush cached writes
In barebox, the dirty block cache for a given device is flushed when
closing all file descriptors to it, a cached block needs to be evicted
due to pressure or when barebox shuts down.

There are valid uses to do it explicitly however:

  - When benchmarking write speed to a block device
  - Before forced resets, because a regular reset would run shutdown
    code that should be omitted.

Add a sync command that does this. Unlike the userspace variant, we
don't bother to implement a argument-less sync that's global. Users are
supposed to flush the files they are actually interested in, e.g.:

  sync /dev/mmc2.0

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240731080510.364706-10-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>