console: honour OF current-speed property if available
current-speed describes the current baud rate the device operates at.
It's not used often in DT, but we should honour it if available.

This gives boards the ability to configure consoles that are activated
early (e.g. ACTIVATE_ALL) with the correct baud rate from the beginning
instead of activating it at CONFIG_BAUDRATE and then switching the
baudrate at environment load time.

Additionally, this can be used to work around following bug:

 - Console is added, but not activated automatically
 - NV sets cs0.active="ioe" cs0.baudrate="4000000"
 - .active parameter is applied first, so CONFIG_BAUDRATE is configured
 - .baudrate parameter is applied second, so baudrate switch happens
 - System hangs until Enter is pressed:
   ## Switch baudrate on console serial1 to 4000000 bps and press ENTER

until it's resolved in a cleaner manner.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240801070918.1667337-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>